requirements.txt
# Portfolio Backtest System Dependencies

# Core data processing
pandas>=1.3.0
numpy>=1.20.0

# Visualization 
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.8.0

# Dashboard
dash>=2.6.0
dash-bootstrap-components>=1.2.0

# Data analysis
scipy>=1.7.0

# System utilities
python-dateutil>=2.8.0

# Deployment
gunicorn>=20.1.0
Pillow>=9.0.0  # For image processing 